pulled out testinitialize
