+++ playwright/accounts.spec.js
@@ -1,31 +1,8 @@
 const { test, expect } = require("@playwright/test");
-const { _electron: electron } = require("playwright");
-const rimraf = require("rimraf");
-const fs = require("fs");
-const path = require("path");
-
-let electronApp;
-
-const initialize = async userData => {
-  const userDataPath = path.join(__dirname, "/temp");
-  if (fs.existsSync(`${userDataPath}`)) {
-    rimraf.sync(userDataPath);
-  }
-
-  fs.mkdirSync(userDataPath, { recursive: true });
-
-  if (userData) {
-    const jsonFile = path.resolve("playwright/setups/", `${userData}.json`);
-    fs.copyFileSync(jsonFile, `${userDataPath}/app.json`);
-  }
-
-  electronApp = await electron.launch({
-    args: ["./.webpack/main.bundle.js", `--user-data-dir=${userDataPath}`],
-  });
-};
+const initialize = require("./testInitialize");
 
 test("get the app running", async () => {
-  await initialize("1AccountBTC1AccountETH");
+  const electronApp = await initialize("1AccountBTC1AccountETH");
   const window = await electronApp.firstWindow();
 
   const title = await window.title();
@@ -33,7 +10,7 @@ test("get the app running", async () => {
 
   await window.screenshot({ path: "start-of-test.png" });
 
-  await window.pause();
+  // await window.pause();
   await window.screenshot({ path: "end-of-test.png" });
 
   await electronApp.close();
